+++ tests/testthat/test-explain.R
@@ -17,7 +17,8 @@ test_that("explain provides correct result with data.frame", {
                     batch_size = 512,
                     attention_width = 1,
                     num_shared = 1,
-                    num_independent = 1)
+                    num_independent = 1,
+                    verbose = FALSE)
 
   expect_equal(which.max(fit$fit$importances$importance), 1)
   expect_equal(fit$fit$importances$variables, colnames(x))
@@ -53,13 +54,15 @@ test_that("explain works for dataframe, formula and recipe", {
 
   # formula
   tabnet_pretrain <- tabnet_pretrain(Sale_Price ~., data=small_ames, epochs = 3, valid_split=.2,
-                                     num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1)
+                                     num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1,
+                                     verbose = FALSE)
   expect_no_error(
     tabnet_explain(tabnet_pretrain, new_data=small_ames)
   )
 
   tabnet_fit <- tabnet_fit(Sale_Price ~., data=small_ames, tabnet_model=tabnet_pretrain, epochs = 3,
-                           num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1)
+                           num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1,
+                           verbose = FALSE)
   expect_no_error(
     tabnet_explain(tabnet_fit, new_data=small_ames)
   )
@@ -70,13 +73,15 @@ test_that("explain works for dataframe, formula and recipe", {
     step_normalize(all_numeric_predictors())
 
   tabnet_pretrain <- tabnet_pretrain(rec, data=small_ames, epochs = 3, valid_split=.2,
-                                     num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1)
+                                     num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1,
+                                     verbose = FALSE)
   expect_no_error(
     tabnet_explain(tabnet_pretrain, new_data=small_ames)
   )
 
   tabnet_fit <- tabnet_fit(rec, data=small_ames, tabnet_model=tabnet_pretrain, epochs = 3,
-                           num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1)
+                           num_steps = 1, attention_width = 1, num_shared = 1, num_independent = 1,
+                           verbose = FALSE)
   expect_no_error(
     tabnet_explain(tabnet_fit, new_data=small_ames)
   )
@@ -100,14 +105,16 @@ test_that("support for vip on tabnet_fit and tabnet_pretrain", {
                     batch_size = 512,
                     attention_width = 1,
                     num_shared = 1,
-                    num_independent = 1)
+                    num_independent = 1,
+                    verbose = FALSE)
 
   fit <- tabnet_fit(x, y, epochs = 1,
                     num_steps = 1,
                     batch_size = 512,
                     attention_width = 1,
                     num_shared = 1,
-                    num_independent = 1)
+                    num_independent = 1,
+                    verbose = FALSE)
 
   expect_no_error(vip::vip(pretrain))
   expect_no_error(vip::vip(fit))
@@ -135,7 +142,8 @@ test_that("Importance is skipped if skip_importance flag is used", {
                     attention_width = 1,
                     num_shared = 1,
                     num_independent = 1,
-                    skip_importance = TRUE)
+                    skip_importance = TRUE,
+                    verbose = FALSE)
 
   expect_equal(fit$fit$importances, NULL)
 
@@ -145,7 +153,8 @@ test_that("Importance is skipped if skip_importance flag is used", {
                     attention_width = 1,
                     num_shared = 1,
                     num_independent = 1,
-                    skip_importance = FALSE)
+                    skip_importance = FALSE,
+                    verbose = FALSE)
 
 
   expect_equal(which.max(fit$fit$importances$importance), 1)
@@ -169,7 +178,8 @@ test_that("explain works for parsnip model", {
 
 test_that("explain works for multi-outcome classification model", {
 
-  fit <- tabnet_fit(x, data.frame(y = y, z = y + 1), epochs = 1)
+  fit <- tabnet_fit(x, data.frame(y = y, z = y + 1), epochs = 1,
+                    verbose = FALSE)
 
   expect_no_error(tabnet_explain(fit, new_data = x))